+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Issue #807: transparency setting for child layers is getting set/initialized now for grouped layers.
 - Print: Polygon representations are now correctly shown in the map and in the legend.
 - layerSequence: fix possibility to move layers in tree despite defined layerSequence
+- Print: fix handling for highresolution plotService
 
 ---